Always run alembic migrations during startup
Before this change, starting the backend with a completely empty
database would fail, because initializing all the services would require
an already migrated database, and we would be running migrations only
after initializing the services.
With this, alembic may be invoked multiple times during startup as
uvicorn will start workers in batches, but that should not cause any
problems.