Fix login screen flickering
When checking whether the user is logged in, the auth provider would set
the login status to "logged in", although in truth the user was neither
logged in nor logged out yet. We now have a third "loading" login status
that allows components to precisely control what to render in which
moments during the login process. Crucially, the login screen now only
redirects to the dashboard if the user is definitely logged in, removing
the flickering we've had before.